+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `deep_remove` now removes the last existing key in the path, ignoring the rest
+  to keep it in line with the behavior of `maps:remove/1` which silently returns
+  the map if the key doesn't exist.
+
 ## [0.3.0] - 2018-09-13
 
 Initial release